extracting the inner carousel fragment classes
